Move static helpers from MapSlots to SlotUtils
